// sniperbot/src/types.ts
export interface CurrentTransaction {
  token: string;
  tokenTime: number;
  purchasePrice: number;
  balance: number;
  solPaid: number;
  solPaidUSDC: number;
  baseVault?: string;
  quoteVault?: string;
  poolAddress?: string;
}